feat: better progress bar and download tracking

+++ MainForm.cs
@@ -1024,7 +1024,7 @@ namespace AndroidSideloader
             if (fileDialogResult == DialogResult.OK)
             {
                 string selectedPath = fileDialog.FileName;
-                Console.WriteLine("Selected .ab file: " + selectedPath);
+                Logger.Log("Selected .ab file: " + selectedPath);
 
                 _ = FlexibleMessageBox.Show(Program.form, "Click OK on this Message...\r\nThen on your Quest, Unlock your device and confirm the backup operation by clicking on 'Restore My Data'\r\nRookie will remain frozen until the process is completed.");
                 output_abRestore = ADB.RunAdbCommandToString($"adb restore \"{selectedPath}").Output;
@@ -1037,7 +1037,7 @@ namespace AndroidSideloader
             if (folderDialogResult == DialogResult.OK)
             {
                 string selectedFolder = folderDialog.SelectedPath;
-                Console.WriteLine("Selected folder: " + selectedFolder);
+                Logger.Log("Selected folder: " + selectedFolder);
 
                 Thread t1 = new Thread(() =>
                 {
@@ -1838,7 +1838,7 @@ namespace AndroidSideloader
                                             ulong releaseGameVersionCode = ulong.Parse(Utilities.StringUtilities.KeepOnlyNumbers(releaseGame[SideloaderRCLONE.VersionCodeIndex]));
                                             if (releaseGameVersionCode > cloudVersionInt)
                                             {
-                                                Console.WriteLine($"Updated cloudVersionInt for {packagename} from {cloudVersionInt} to {releaseGameVersionCode}");
+                                                Logger.Log($"Updated cloudVersionInt for {packagename} from {cloudVersionInt} to {releaseGameVersionCode}");
                                                 cloudVersionInt = releaseGameVersionCode;
                                             }
                                         }
@@ -2644,6 +2644,7 @@ Things you can try:
                     string packagename = Sideloader.gameNameToPackageName(gameName);
                     string dir = Path.GetDirectoryName(gameName);
                     string gameDirectory = Path.Combine(Properties.Settings.Default.downloadDir, gameName);
+                    string downloadDirectory = Path.Combine(Properties.Settings.Default.downloadDir, gameName);
                     string path = gameDirectory;
 
                     string gameNameHash = string.Empty;
@@ -2693,11 +2694,12 @@ Things you can try:
 
                         if (doDownload)
                         {
+                            downloadDirectory = $"{Properties.Settings.Default.downloadDir}\\{gameNameHash}";
                             _ = Logger.Log($"rclone copy \"Public:{SideloaderRCLONE.RcloneGamesFolder}/{gameName}\"");
                             t1 = new Thread(() =>
                             {
                                 string rclonecommand =
-                                $"copy \":http:/{gameNameHash}/\" \"{Properties.Settings.Default.downloadDir}\\{gameNameHash}\" {extraArgs} --progress --rc";
+                                $"copy \":http:/{gameNameHash}/\" \"{downloadDirectory}\" {extraArgs} --progress --rc --check-first --fast-list";
                                 gameDownloadOutput = RCLONE.runRcloneCommand_PublicConfig(rclonecommand);
                             });
                         }
@@ -2709,19 +2711,30 @@ Things you can try:
                     else
                     {
                         _ = Directory.CreateDirectory(gameDirectory);
-                        _ = Logger.Log($"rclone copy \"{currentRemote}:{SideloaderRCLONE.RcloneGamesFolder}/{gameName}\"");
+                        downloadDirectory = $"{SideloaderRCLONE.RcloneGamesFolder}/{gameName}";
+                        _ = Logger.Log($"rclone copy \"{currentRemote}:{downloadDirectory}\"");
                         t1 = new Thread(() =>
                         {
-                            gameDownloadOutput = RCLONE.runRcloneCommand_DownloadConfig($"copy \"{currentRemote}:{SideloaderRCLONE.RcloneGamesFolder}/{gameName}\" \"{Properties.Settings.Default.downloadDir}\\{gameName}\" {extraArgs} --progress --rc --retries 1 --low-level-retries 1");
+                            gameDownloadOutput = RCLONE.runRcloneCommand_DownloadConfig($"copy \"{currentRemote}:{downloadDirectory}\" \"{Properties.Settings.Default.downloadDir}\\{gameName}\" {extraArgs} --progress --rc --retries 1 --low-level-retries 1 --check-first");
                         });
+                    }
+
+                    if (Directory.Exists(downloadDirectory)) {
+                        string[] partialFiles = Directory.GetFiles($"{downloadDirectory}", "*.partial");
+                        foreach (string file in partialFiles)
+                        {
+                            File.Delete(file);
+                            _ = Logger.Log($"Deleted partial file: {file}");
+                        }
                     }
 
                     t1.IsBackground = true;
                     t1.Start();
 
                     changeTitle("Downloading game " + gameName, false);
+                    speedLabel.Text = "Starting download...";
+                    etaLabel.Text = "Please wait...";
 
-                    int i = 0;
                     //Download
                     while (t1.IsAlive)
                     {
@@ -2729,42 +2742,58 @@ Things you can try:
                         {
                             HttpResponseMessage response = await client.PostAsync("http://127.0.0.1:5572/core/stats", null);
                             string foo = await response.Content.ReadAsStringAsync();
-                            Debug.WriteLine("RESP CONTENT " + foo);
+                            //Debug.WriteLine("RESP CONTENT " + foo);
                             dynamic results = JsonConvert.DeserializeObject<dynamic>(foo);
 
                             if (results["transferring"] != null)
                             {
-                                long allSize = 0;
-                                long downloaded = 0;
+                                double totalSize = 0;
+                                double downloadedSize = 0;
+                                long fileCount = 0;
+                                long transfersComplete = 0;
+                                long totalChecks = 0;
+                                long globalEta = 0;
+                                float speed = 0;
+                                float downloadSpeed = 0;
+                                double estimatedFileCount = 0;
 
-                                foreach (dynamic obj in results.transferring)
+                                totalSize = results["totalBytes"];
+                                downloadedSize = results["bytes"];
+                                fileCount = results["totalTransfers"];
+                                totalChecks = results["totalChecks"];
+                                transfersComplete = results["transfers"];
+                                globalEta = results["eta"];
+                                speed = results["speed"];
+                                estimatedFileCount = Math.Ceiling(totalSize / 524288000); // maximum part size
+
+                                if (totalChecks > fileCount)
                                 {
-                                    allSize += obj["size"].ToObject<long>();
-                                    downloaded += obj["bytes"].ToObject<long>();
+                                    fileCount = totalChecks;
+                                }
+                                if (estimatedFileCount > fileCount)
+                                {
+                                    fileCount = (long)estimatedFileCount;
                                 }
 
-                                float downloadSpeed = results.speed.ToObject<float>() / 1000000;
-                                allSize /= 1000000;
-                                downloaded /= 1000000;
+                                downloadSpeed = speed / 1000000;
+                                totalSize /= 1000000;
+                                downloadedSize /= 1000000;
 
-                                Debug.WriteLine("Allsize: " + allSize + "\nDownloaded: " + downloaded + "\nValue: " + (downloaded / (double)allSize * 100));
+                                // Logger.Log("Files: " + transfersComplete.ToString() + "/" + fileCount.ToString() + " (" + Convert.ToInt32((downloadedSize / totalSize) * 100).ToString() + "% Complete)");
+                                // Logger.Log("Downloaded: " + downloadedSize.ToString() + " of " + totalSize.ToString());
 
                                 progressBar.Style = ProgressBarStyle.Continuous;
-                                progressBar.Value = Convert.ToInt32(downloaded / (double)allSize * 100);
+                                progressBar.Value = Convert.ToInt32((downloadedSize / totalSize) * 100);
 
-                                i++;
-                                if (i == 4)
-                                {
-                                    i = 0;
-                                    float seconds = (allSize - downloaded) / downloadSpeed;
-                                    TimeSpan time = TimeSpan.FromSeconds(seconds);
-                                    etaLabel.Text = "ETA: " + time.ToString(@"hh\:mm\:ss") + " left";
-                                }
+                                TimeSpan time = TimeSpan.FromSeconds(globalEta);
+                                etaLabel.Text = etaLabel.Text = "ETA: " + time.ToString(@"hh\:mm\:ss") + " left";
 
-                                speedLabel.Text = "DLS: " + string.Format("{0:0.00}", downloadSpeed) + " MB/s";
+                                speedLabel.Text = "DLS: " + transfersComplete.ToString() + "/" + fileCount.ToString() + " files - " + string.Format("{0:0.00}", downloadSpeed) + " MB/s";
                             }
                         }
-                        catch { }
+                        catch
+                        {
+                        }
 
                         await Task.Delay(100);
 
@@ -2772,25 +2801,34 @@ Things you can try:
 
                     if (removedownloading)
                     {
-                        changeTitle("Deleting game files", false);
+                        changeTitle("Keep game files?", false);
                         try
                         {
                             cleanupActiveDownloadStatus();
-                            if (hasPublicConfig)
-                            {
-                                if (Directory.Exists($"{Properties.Settings.Default.downloadDir}\\{gameNameHash}"))
-                                {
-                                    Directory.Delete($"{Properties.Settings.Default.downloadDir}\\{gameNameHash}", true);
-                                }
 
-                                if (Directory.Exists($"{Properties.Settings.Default.downloadDir}\\{gameName}"))
-                                {
-                                    Directory.Delete($"{Properties.Settings.Default.downloadDir}\\{gameName}", true);
-                                }
-                            }
-                            else
+                            DialogResult res = FlexibleMessageBox.Show(
+                                $"{gameName} already has some downloaded files, do you want to delete them?\n\nClick NO to keep the files if you wish to resume your download later.",
+                                "Delete Temporary Files?", MessageBoxButtons.YesNo);
+
+                            if (res == DialogResult.Yes)
                             {
-                                Directory.Delete(Properties.Settings.Default.downloadDir + "\\" + gameName, true);
+                                changeTitle("Deleting game files", false);
+                                if (hasPublicConfig)
+                                {
+                                    if (Directory.Exists($"{Properties.Settings.Default.downloadDir}\\{gameNameHash}"))
+                                    {
+                                        Directory.Delete($"{Properties.Settings.Default.downloadDir}\\{gameNameHash}", true);
+                                    }
+
+                                    if (Directory.Exists($"{Properties.Settings.Default.downloadDir}\\{gameName}"))
+                                    {
+                                        Directory.Delete($"{Properties.Settings.Default.downloadDir}\\{gameName}", true);
+                                    }
+                                }
+                                else
+                                {
+                                    Directory.Delete(Properties.Settings.Default.downloadDir + "\\" + gameName, true);
+                                }
                             }
                         }
                         catch (Exception ex)
